refactor: jdbc RowMapper 관리지점 통일
