+++ src/main/java/roomescape/repository/ThemeJdbcRepository.java
@@ -5,22 +5,12 @@ import java.util.Map;
 import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
 import roomescape.model.Theme;
 
 @Repository
 public class ThemeJdbcRepository implements ThemeRepository {
-
-    static final RowMapper<Theme> THEME_ROW_MAPPER =
-        (rs, rowNum) -> {
-            var id = rs.getLong("id");
-            var name = rs.getString("name");
-            var description = rs.getString("description");
-            var thumbnail = rs.getString("thumbnail");
-            return new Theme(id, name, description, thumbnail);
-        };
 
     private final JdbcTemplate jdbcTemplate;
 
@@ -35,7 +25,7 @@ public class ThemeJdbcRepository implements ThemeRepository {
             where T.id = ?
             """;
 
-        var themeList = jdbcTemplate.query(sql, THEME_ROW_MAPPER, id);
+        var themeList = jdbcTemplate.query(sql, RowMappers.THEME, id);
         return themeList.stream().findAny();
     }
 
@@ -63,6 +53,6 @@ public class ThemeJdbcRepository implements ThemeRepository {
             select T.id, T.name, T.description, T.thumbnail from THEME T
             """;
 
-        return jdbcTemplate.query(sql, THEME_ROW_MAPPER);
+        return jdbcTemplate.query(sql, RowMappers.THEME);
     }
 }
